refactor : 테스트코드 수정 - 테스트 컨테이너에서 데이터베이스를 생성하고 테스트 코드 수행할 수 있도록 코드 수정
