+++ src/test/java/com/ssafy/ssafy_project/support/TestInfraConfig.java
@@ -1,0 +1,57 @@
+package com.ssafy.ssafy_project.support;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.ssafy.ssafy_project.global.application.port.out.RefreshTokenPortOut;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Primary;
+import org.testcontainers.containers.PostgreSQLContainer;
+
+import java.util.Map;
+import java.util.Optional;
+import java.util.concurrent.ConcurrentHashMap;
+
+@TestConfiguration(proxyBeanMethods = false)
+public class TestInfraConfig {
+
+    @Bean
+    @ServiceConnection
+    PostgreSQLContainer<?> postgresContainer() {
+        return new PostgreSQLContainer<>("postgres:16-alpine");
+    }
+
+    @Bean
+    @Primary
+    RefreshTokenPortOut refreshTokenPortOut() {
+        return new InMemoryRefreshTokenPortOut();
+    }
+
+    @Bean
+    ObjectMapper objectMapper() {
+        return new ObjectMapper();
+    }
+
+    public static class InMemoryRefreshTokenPortOut implements RefreshTokenPortOut {
+        private final Map<Long, String> storage = new ConcurrentHashMap<>();
+
+        @Override
+        public void save(Long userId, String refreshToken, long expirationMillis) {
+            storage.put(userId, refreshToken);
+        }
+
+        @Override
+        public Optional<String> find(Long userId) {
+            return Optional.ofNullable(storage.get(userId));
+        }
+
+        @Override
+        public void delete(Long userId) {
+            storage.remove(userId);
+        }
+
+        public void clear() {
+            storage.clear();
+        }
+    }
+}
